Tests for Check.py. Other changes:
Reimplementing most of the test utilities in a nicer way - had waaaay too many function names appearing.

All tests for whether an error is raised now match for an expected error string.

Stopped allowing funding sum and modpoint of empty sequence of coordinates, and added tests for this.

Added a dimension property to Paulis that just echoes the Qubit's dimension. Added tests for this.

Added extra code in PauliProduct to ensure coordinates all valid. Added tests for this.