Tests for Detector.py. Other changes:
Removing restrictions on what Paulis can make up a check (previously was only sign 1 and XYZ letters). It should instead be up to a syndrome extractor to complain that it doesn't know how to handle a Pauli, if that's the case.

Instead, replacing it with a test for whether Paulis in a check have a Hermitian product. Added tests for this.

Adding 'has_tuple_coords' property to a few classes, and tests for this.